Added several sections to installation chapter of user's guide.
This section describes the various ways of installing RPMD. I
considered putting it in with the introduction, but I thought it was
important enough to be worth its own chapter, if for no other reason
than to make it easier to find. Again, this should be revised and
expounded upon as the codebase grows.