Added Fortran module for representing transition state dividing surface.
The transition_state module defines a dividing surface for a reactive
transition state, defined by a set of forming bonds and a set of
breaking bonds. The module is also designed to handle multiple
equivalent transition states. Subroutines are available for evaluating
the value, gradient, and Hessian of the transition state dividing
surface. The use of a Fortran module greatly simplifies the parameter
lists of these subroutines, as we don't need to pass the transition
state attributes as parameters to these subroutines.