Added TransitionState class to new rpmd.surface module.
This class is designed to contain information needed to represent the
dividing surface between reactants and products, an important input
parameter to an RPMD rate calculation. In particular, RPMD simulations
require not just the dividing surface function, but also its gradient
and Hessian. This class is equipped to handle multiple equivalent
transition states.

The TransitionState class implements a somewhat generalized dividing
surface based on the lengths of certain important bonds that are forming
or breaking due to the reaction. You provide a transition state structure
and a list of the bonds being broken and formed. The implementation is
based on the H + CH4 dividing surface function be Suleimanov et al, and
even allows for handling of multiple equivalent transition states as
done in that paper. I think this should work for at least hydrogen
abstraction reactions, and perhaps other reaction families as well.

I suspect that we could evaluate the gradient and Hessian numerically
for really complicated dividing surfaces, but I think I managed to
obtain the analytical versions for this function, so we'll use them.