Added Fortran random number subroutines.
The random() and randomn() subroutines return random numbers from a
uniform and normal distribution, respectively. They are designed so that
we can easily change the pseudo-random algorithm if desired; for now we
are using the random_number() intrinsic function.

The random_init() subroutine is used to initialize (seed) the
pseudo-random number generator. The implementation is based on an
example in the gfortran manual, and uses the intrinsic random_seed()
subroutine to seed based on the system clock.