Added first implementation of RPMD trajectory evolution algorithm.
The verlet_step() subroutine is used to take a single time step using
the velocity Verlet algorithm to integrate the classical Hamiltonian
for the ring polymer system. It turns out that the ring polymer term in
the Hamiltonian is best handled in normal mode space (see
doi:10.1063/1.3489925 for more); this is implemented in the
free_ring_polymer_step() subroutine.

The Hamiltonian also requires the potential and forces of the system.
The verlet_step() subroutine requires a parameter that is used to call
back to Python to evaluate a user-provided function for evaluating
the potential and forces. This is done purposefully, to enable the user
to use a quantum chemistry package to get energies if a mapped PES is
not available.