When activating a dividing surface, optionally specify the module to use.
It turns out that the rpmd._main module has its own copy of the dividing
surface modules defined in rpmd/_system.f90, so we need to set the
values in that module in order for the RPMD simulation to be able to see
them. To do this with the most flexibility, we simply update the
activate() methods to accept as an optional parameter the module to set
the parameters in. This will use the rpmd._system module by default.