Added equilibrate() subroutine for equilibrating an RPMD trajectory.
This subroutine carries out an RPMD trajectory simulation in the
presence of an Andersen thermostat, with option to constrain the
centroids to the transition state dividing surface. No physical
quantities are computed during this trajectory evolution; it is
generally used as a randomizer for other trajectories over which a
physical quantity is computed.