Added recrossing_trajectory() for computing the recrossing factor.
This subroutine conducts an RPMD trajectory simulation, updating the
value of the recrossing factor at each time step. Running this
subroutine for a large sampling of trajectories should cause the
recrossing factor to converge. If your choice of transition state
dividing surface is close to optimal, the value of the recrossing factor
will be close to one.