Added Python implementation of recrossing factor calculation.
The high-level logic of the recrossing factor calculation is now written
in Python. This aids in clarity while not introducing a significant speed
penalty, as the bottleneck in the calculation remains in the RPMD
trajectory evolutions. Furthermore, we may be able to parallelize the
algorithm in the future if desired.