Added subroutines for including umbrella potential and bias term.
These terms are only included if mode is set to 1. The umbrella potential
comes with a new parameter, kforce, representing the force constant of
the potential. This parameter is now included in the system Fortran
module.

These terms will be needed for computing the static factor, part of the
procedure of using RPMD to compute reaction rate coefficients.