First multiprocessor version of RPMD umbrella integration algorithm.
In the new algorithm, once each configuration has been equilibrated in
each window, a number of jobs are asynchronously spawned for sampling
trajectories in each window before continuing to equilibrate in the
next window. These jobs are spawned to a pool of processors, which
achieves a good load balancing, as sampling trajectories will likely
collect in a queue, since the equilibrations are faster. Note that
at the moment it makes a complete copy of the memory for each subprocess,
which is probably okay since we're not at all memory limited.

Also note that, for this to work, your RPMD object must be picklable. In
particular, that means that a pure Python module-level function must be
stored in the RPMD.potential attribute. This can simply just call an
f2py-wrapped function, however; this doesn't seem to be much of an
additional speed penalty in single-process jobs, and is therefore worth
it to enable multiprocess support.