First multiprocessor version of RPMD recrossing factor algorithm.
In the new algorithm, the child trajectories sampled as the parent
trajectory evolves are spawned to a pool of processors. Since the child
trajectories are short compared to the parent trajectory, the benefit
of multiple processors is much smaller than in the umbrella integration
algorithm. Nonetheless, there is somewhat of a benefit, so I think it's
worth implementing. As before, a complete copy of the memory is made for
each subprocess, and you must make sure your RPMD object is picklable by
wrapping any f2py-wrapped potential function in a pure Python function
defined at module level.