Only use one bead when generating initial umbrella positions.
Since we are going to equilibrate in each window anyway, we don't need to
include all of the beads when generating the initial positions. This
suggests that the initial position generation will be very fast compared
to the sampling in each window (when many beads are used), so I've
moved the spawning of sampling trajectories to occur after initial
positions are generated in all windows.