Allow user to specify different umbrella force constant in each window.
The optimum value of the force constant depends on the position along the
reaction coordinate. For instance, farther from xi = 1 you can use a
smaller force constant to improve the convergence of the umbrella
sampling in that window. This commit allows for either a single force
constant or an array of force constants (one per window) to be specified.