Specify atom symbols instead of masses for Reactants.
A new module rpmd.element was added to store a table of the atomic masses
for common isotopes of the first forty elements of the periodic table.
When a Reactants object is created, this dict is used to convert the
atomic symbols to masses automatically, saving the user the trouble of
doing so by hand.