TransitionState class now represents a single transition state.
The previous implementation was designed to be able to store multiple
transition states, such as the four equivalent transition states of
H + CH4 -> H2 + CH3. Instead, we now store these as separate
TransitionState objects in a list on the RPMD object. This requires a
custom procedure for activating in the Fortran layer (i.e. we cannot use
the TransitionState.activate() method).