Added handling of units to various methods.
This makes these methods more suitable for use in an RPMD "input file",
which is currently envisioned as a simple Python script. The unit
conversion functions recently defined in the rpmd.quantity module are
used.

Also updated the unit tests of the Reactants and TransitionState class
to reflect the use of these units.