Added option for specifying thermostat at Fortran level.
The desired thermostat is indicated by the value of the thermostat
module-level variable (e.g. 1 = Andersen). Other module-level variables
are used to store parameters for each thermostat, e.g. the Andersen
thermostat sampling time.

This is in preparation for allowing the user to choose from multiple
thermostats. At the moment only the Andersen thermostat is available.