Added representation of Andersen thermostat to Python layer.
The new rpmd.thermostat module is the place to store representations of
various thermostats. Currently the only available thermostat is the
Andersen thermostat.

The thermostat is now a required parameter to both the recrossing factor
and umbrella integration calculations. You should pass an instance of a
thermostat class from the rpmd.thermostat module.