Added RPMD.generateUmbrellaConfigurations() method.
This method is used to generate a set of initial configurations along
the reaction coordinate for future use in RPMD umbrella sampling. The
algorithm starts near the transition state dividing surface and moves
away in either direction, running an RPMD equilibration in each window
to obtain the appropriate configuration in that window. That
configuration is then used as the initial position for determining the
configuration in the next window.