Get computeStaticFactor() geometries from generateUmbrellaConfigurations().
The code that generated the initial geometries has now been removed from
computeStaticFactor(), since it is now in generateUmbrellaConfigurations().
The calculation of the initial geometries for umbrella sampling is
sufficiently different than the sampling in each window that it is worth
having its own separate function. computeStaticFactor() now raises an
exception if generateUmbrellaConfigurations() has not been run for the
system.

Note that you don't necessarily need to use the same xi_list and kforce
for computeStaticFactor() as you would for umbrella sampling, although it
is probably good practice to do so.