Added Window class for representing an umbrella integration window.
This class enables the user to specify different parameters for umbrella
sampling for each window if desired. It also provides a place to store
the mean and variance of the reaction coordinate in each window.

The computeStaticFactor() method has been updated to use a list of Window
objects instead of separate parameters (xi_list et al.). This makes the
method somewhat easier to read, though it does potentially complicate the
input file a bit. The tradeoff is probably worth it, though.