Temperature now set when creating RPMD object.
The sequence of method calls that lead to an RPMD rate coefficient
require that intermediate data be stored on the RPMD object. This
intermediate data is bound to the temperature of interest, which must be
the same throughout the sequence of method calls. For this reason, it
makes sense to simply make the temperature an attribute of the RPMD
object and set it once (at initialization time).