Enable seeding of random number generator at Python level.
It is very important to get the best possible pseudo-random number
generation in order to get good RPMD sampling statistics. It is equally
incorrect to seed the random number generator too frequently as it is to
not do so frequently enough. The current practice is to seed the random
number generator once at each step of the RPMD calculation, so that each
step could also be run independently if desired. Enabling this requires
exposing the Fortran random_init() subroutine to Python.