Fixed bug in transition state dividing surface.
When we changed the TransitionState class to represent a single dividing
surface (and therefore require multiple TransitionState objects to
represent multiple equivalent dividing surfaces), we forgot to update
which dimension of the formingBonds and breakingBonds attributes reflects
the number of each. For the H + CH4 example, which has one of each, this
was resulting in double-counting, since it was using two instead of one.
The result was slightly incorrect values of xi due to s1 being twice
what it should have been. This commit fixes the error.