Fixes to the umbrella integration algorithm.
There were two mistakes: a misinterpretation of how to assemble N (the
number of sampling points in each window during the umbrella sampling,
not the number of bins per window) and incorrectly computing the mean and
variance from <xi> and <xi^2> in each bin. Fixing these issues results in
our being able to correctly reproduce a potential of mean force
calculation for H + CH4.