Modified umbrella sampling algorithm to ensure better statistics.
The key idea here is that we want to give each umbrella sampling
trajectory a unique initial position and momentum to get the best
possible statistics. To do this, I have moved the equilibration in each
window to always be done in the main process, using it to spawn sampling
trajectories in much the same way as in the recrossing factor algorithm.
This is a small tradeoff in parallelization, but seems to be worth it to
ensure better statistical sampling.

I don't think we need to do the full equilibration between each spawned
sampling algorithm, as long as the time evolution is long enough for
several applications of the Andersen thermostat. Perhaps in the future
we could add an option to specify this independently.