Added RPMD.saveRecrossingFactor() for saving recrossing factor results.
The file is saved in a subdirectory named for the current temperature of
interest. The file name includes the number of beads, so that we could
run both classical and RPMD trajectories in the same location. (Are there
any other parameters we ought to use to mark distinct jobs?)

To achieve this, an attribute was added to the RPMD object for storing
the output directory. A method createWorkingDirectory() was also added
for creating the output directory on disk if needed. By default the
output directory is placed in the same directory the job is executed
from.