Added RPMD.loadRecrossingFactor() for loading recrossing factor results.
If a recrossing factor results file exists for the given temperature
and number of beads, and if it uses the same parameters as the requested
calculation (particularly xi_current), then the saved results will be
included, as if restarting from that point. This provides an easy way of
adding additional trajectories if you don't feel like you have good
convergence, without having to repeat the existing trajectories first.

The flow of the RPMD.computeTransmissionCoefficient() method was
restructured a bit to accomodate the new restart functionality.