Added functionality for saving umbrella sampling results.
This is not placed in a separate method so that we can save the results
incrementally as trajectories for each window are completed. This way we
can still restart even if the job is interrupted and retain at least some
of the progress.

We need to save the accumulated values of <xi> and <xi^2> and the number
of sampling points in order to restart. Note that, in order to recover
the variance, we need to print many decimal places for <xi> and <xi^2>.