Added RPMD.loadUmbrellaSampling() for loading umbrella sampling results.
If the reaction coordinate of the current job is the same as the saved
job, the trajectories are not repeated unless additional sampling points
are requested. This is applied on a per-window basis, so individual
windows might be resampled if they do not match any windows in the saved
output.

In order to be consistent with the algorithm flow when no saved output
exists, we still go through the process of spawning jobs to the pool of
processes. However, these processes have evolution times of zero, so they
complete instantaneously.