Added RPMD.saveRateCoefficient() for saving rate coefficient results.
This is again a very quick calculation, but is the ultimate result we are
after (at least in a rate calculation), so it's very important that we
save the results properly. Note that k(T) values are given in both
cm^3/molecule*s and cm^3/mol*s to satisfy both the scientists and the
engineers.