New method for specifying multiple equivalent transition states.
The new method only requires that the geometry be specified for the first
transition state. The idea is to specify the first transition state as a
TransitionState object in the RPMD.__init__() method, and specify the
others via the new addEquivalentTransitionState() method. The important
thing is that the bonds are in the same order and with consistent
indices, so that the correct bond lengths are used in the equivalent
transition states.