Added subdirectories for various numbers of beads for output files.
This is in anticipation of saving many more files for each temperature
and number of beads, particularly in the umbrella sampling. For now,
I have changed the Nbeads to be a parameter set in the RPMD __init__()
method instead of for each of the umbrella sampling and recrossing factor
calculations. In the future this may need to change, but it's okay for
now.