New breadth-first, tolerance-based algorithm for umbrella sampling.
The idea is that we would rather get some data in all windows than get
lots of data in a few windows. This enables us to estimate the k(T)
value even if the calculation has not yet converged. It potentially also
allows the algorithm to decide what portion of the reaction coordinate is
worth more computational time.

The new algorithm also introduces a tolerance for automatic termination
of the potential of mean force algorithm. This helps ensure convergence
to the desired tolerance.

Finally, the conductUmbrellaSampling() and computePotentialOfMeanForce()
methods have been merged together from the point of view of the input
file. Update your input files accordingly.