Added tolerance parameter to recrossing factor calculation.
This enables the code to automatically halt the recrossing factor
calculation if the desired tolerance is achieved. Due to the stochastic
nature of the sampling, it is possible for one set of iterations to not
change the recrossing factor much; for this reason, we check the last
several iterations, for which this is far less likely to occur unless
you've actually converged. Still, you may wish to request a slightly
tighter tolerance than you actually want, just to be safe.