Use same xi_current for recrossing factor and rate coefficient calculation.
The RPMD rate coefficient is rigorously independent of the choice of
dividing surface; for this reason it is also independent of the value of
the reaction coordinate chosen to evaluate k(T). For numerical reasons it
is prudent to choose the xi value corresponding to the maximum of the
potential of mean force, but this is not required. In fact, it can be a
good check to choose two different values of xi, and see that the
computed k(T) value is the same for both.

To enable this, a change has been made so that the stored value of
xi used for the recrossing factor calculation is also used to compute the
corresponding k(T). Previously it always used the xi of maximum potential
of mean force.