Don't require multiprocessing to run RPMD calculations.
The multiprocessing package is built into Python 2.6 or later, but must
be installed separately (as a backported package) into earlier versions
of Python. This may be problematic on legacy operating systems. In order
to make our code usable on these systems, we allow the user to run
single-processor jobs even if this package is not installed. The code
will only fail if the user requests a multiprocessor job *and* the
multiprocessing package is not found.

This should also help with debugging.