Now expect info parameter to be returned from potential function.
The info parameter provides a means for the potential function to signal
to the RPMD code that a given geometry is so unphysical that the entire
trajectory should be rejected. This may help avoid sampling of unphysical
regions of the PES. It is currently up to the user implementing the
potential function to devise a means of identifying unphysical geometries;
one way would include checking of certain bond lengths, angles, and/or
dihedrals for extreme values.