Better creation of Fortran-contiguous arrays in recrossing child trajectories.
Previously there was a bug that caused non-Fortran-contiguous arrays to
be created for the recrossing factor child trajectories, especially when
multiple processors are used. This is due to a small detail in the way
the Fortran arrays are created. To be safe, it seems that you should make
the arrays Fortran contiguous as close to the call into the Fortran code
as possible.

The particular bug addressed here involves the creation of the -p child
trajectory. Previously we used -p.copy('F') when in reality we want
(-p).copy('F'); i.e., apply the negation *before* ensuring Fortran
continuity, not *after*.