Fixed type bug in RPMD.computeRateCoefficient().
I'm not entirely sure why, but the value of k_RPMD * fromAtomicUnits
apparently was a str instead of a number, which caused it to be
incompatible with the 'g' format code. I've added a few float() calls to
ensure that k_RPMD * fromAtomicUnits is a float, so that the 'g' format
code can be used.