Include xi_current in recrossing factor and rate coefficient filenames.
In principle the user can choose any value of xi_current to compute the
RPMD rate coefficient at. The value of k_RPMD(T) should be independent of
this choice, as the difference in the static factor should be corrected
for by a similar difference in the recrossing factor. This commit enables
the user to try multiple values of xi_current in the same job, which can
serve as a useful check that the RPMD calculation was successful.