Added Fortran functions for applying GLE thermostat.
Unlike the Andersen thermostat, the GLE thermostat requires a bit of
initialization and cleanup, so I've added the gle_initialize() and
gle_cleanup() subroutines for this purpose. The other new subroutine is
gle_thermostat(), which is used to apply the thermostat itself. The
trajectory subroutines have been modified appropriately to use the GLE
thermostat instead of the Andersen thermostat when desired.