Can now run computeRateCoefficient() without computing recrossing factor.
In many cases the recrossing factor is close to one. For these systems it
is valuable to get an estimate of the rate coefficient without waiting for
the recrossing factor calculation to run. This commit enables a call to
computeRateCoefficient() to succeed even if the recrossing factor has not
been calculated; in this case a value of one will be assumed (and a
warning printed as a reminder).