Added a few sanity checks to the RPMD trajectory codes.
In umbrella integration, we now check that multiple equivalent transition
state dividing surfaces do not give the same maximum value of s1(q).

In umbrella integration, we now check that the breaking and forming bond
lengths do not diverge to crazy geometries (currently 200 bohr), as this
implies that we are sampling way too far from the center of the window.

In recrossing factor, we now check that the breaking and forming bond
lengths do not diverge to crazy geometries (currently 20 bohr), as this
implies that we are sampling way too far from the transition state
dividing surface.