Added rpmdrate.py main execution script with new input file parser.
This commit formalizes and streamlines much of the input process for
running RPMDrate, moving certain steps out of the input file and into the
main execution script. This results in shorter, easier to understand input
files with better error checking of the input information.

Note that the input file is exec()'d as a Python script with builtins
allowed. For this reason it is incumbent upon the end user to be careful
that the input file script he/she is using is not malicious!