Evaluate static factor using potential of mean force at zero as lower bound.
By definition the static factor is exp(A(xi*) - A(0)), where xi = xi* is
at the maximum of A and xi = 0 is at s0(q). (The idea is that this factor
is a correction that shifts a k(T) evaluated at s0 to a k(T) evaluated at
s1. Therefore, you should use the value at xi = 0.