Removed calculation of "old" recrossing factor.
The newer formulation of the recrossing factor expression is
significantly better in that it is guaranteed to approach unity at time
zero and give a slightly better value at long times. The old formulation
was simply included as a check, and is now deleted to avoid confusion on
the part of the end user as to which to use.