Added rpmdrate.interpolate module with several interpolation methods.
This eliminates the dependency on scipy.interpolate, which can be
difficult to install on older platforms. A linear interpolator is now
used instead of a spline for the potential of mean force calculation,
which shouldn't be a problem since many bins can be used to make the
range of interpolation very small.