Attempt to induce more immediate flush of file output in Python.
In principle we shouldn't need to do this, because we want to let the
OS decide how to write output to disk. However, since we are carrying
out rather expensive computations, we want to ensure that the results
are saved immediately so we don't have to repeat them. This is
supposedly done with ``f.flush(); os.fsync(f.fileno())``. I've added
this code snippet in several places to encourage immediate flushing of
written buffers.