Use different formatting code for reaction coordinate in file names.
The "%g" code renders the value of zero differently on various platforms,
which can be annoying when running jobs on different computers. The new
format code "%.4f" should be much more consistent in this regard.
(Hopefully four decimal places is sufficient; this is generally more than
I would consider reporting in a publication.)