Transition state dividing surface can now handle multiple bond pairs.
The transition state dividing surface is now defined as the minimum over
the bonds of the maximum over the equivalent transition states of the
difference between the forming and breaking bond lengths and their
values in the transition state. This formulation implies that, at
s1(q) > 0, all bonds are on the product side of the dividing surface.
This will be useful when working with reactions that involve multiple
pairs of forming and breaking bonds, although I don't have any to test
with at the moment.

Note that this implementation requires that each forming bond be paired
with a breaking bond. Many reaction families with a significant reaction
barrier fall into this category, but not all.