Don't use numpy.trapz in umbrella integration.
Although it claims to do integration using the trapezoid rule, I'm not
sure it does it in exactly the way we want for umbrella integration, so
I've replaced it with a custom implementation. (The trapezoid rule is
easy to implement anyway, so it's not a big deal.) In particular, this
fixes an issue where the potential of mean force for symmetric reactions
was not centered exactly at xi = 1.