Fixed bug in passing of transition state parameters from Python to Fortran.
In the Fortran layer I had combined the number of forming and breaking
bonds into a single parameter since they are currently required to be
equal. This commit also updates the code for passing that information
from Python to Fortran. Otherwise you would either see NaN in your output
or a segfault.