Fixed bug in sampling of momentum while generating umbrella configurations.
The umbrella configurations can be generated with only one bead to save
time, as we only need the centroid positions to launch full umbrella
sampling trajectories. For this reason we wish to override the Nbeads
attribute of the RPMD object (which could have any value) with the local
Nbeads variable (which is always one). However, the sampleMomentum()
method previously was always using the Nbeads attribute, which led to
incorrect momenta being supplied to the umbrella configurations, and
thus to bizarre umbrella configuration geometries. This commit allows
the user to optionally specify an Nbeads parameter to sampleMomentum()
to override the Nbeads attribute, and uses this new feature to get proper
initial momenta for generating umbrella configurations.