Fixed bug in restarting of recrossing factor calculation.
To be able to restart, we need the reaction coordinate maximum xi_current
to be the same in both cases. Previously we were saving that value to
six significant figures, but recently we changed to only using four. As
a result, our check needs to allow for a slightly larger tolerance than
before.