Added xi_range option to Window class for umbrella sampling.
The xi_range attribute indicates the maximum allowed range for the xi value
in a given window. If set and the xi value for a configuration moves outside
of this range, then the trajectory will be restarted. You specify this value
when creating a Window object in your input file.