Recrossing trajectories will now restart if SHAKE fails to converge.
If we start to sample from an unusual region of the PES, we could run into a
situation where the SHAKE algorithm for constraining the trajectory to the
dividing surface is unsuccessful. Now that we have been using the code for a
while, I feel better saying that this is an actual error (rather than a bug
in our code), so when it happens we simply re-equilibrate from the original
configuration and continue.