Added some validation checks to loadUmbrellaSampling().
These check that the saved umbrella sampling trajectories are reasonable, by
checking that the mean is actually within the valid range and the variance
does not jump significantly. If an invalid trajectory is found, a (hopefully
useful) message is printed and the trajectory is rejected.

Invalid umbrella trajectories sometimes occur with potential energy surfaces
that have regions of unphysical poor fit. We also have some checks within
the conductUmbrellaSampling() method; this commit simply adds checks on
already complete trajectories.